refactor: Winner 내부의 로직 Cars로 이동
