+++ src/main/java/racingCar/domains/Cars.java
@@ -13,6 +13,10 @@ public class Cars {
         }
     }
 
+    public Cars(List<Car> cars) {
+        this.cars = cars;
+    }
+
     public void move() {
         for (Car car : this.cars) {
             car.move();
@@ -21,5 +25,30 @@ public class Cars {
 
     public List<Car> getCars() {
         return this.cars;
+    }
+
+    public List<Car> findWinners() {
+        int max = maxPosition();
+        return winners(max);
+    }
+
+    private int maxPosition() {
+        int max = 0;
+        for (Car car : this.cars) {
+            if (car.getPosition() > max) {
+                max = car.getPosition();
+            }
+        }
+        return max;
+    }
+
+    private List<Car> winners(int max) {
+        List<Car> winners = new ArrayList<>();
+        for (Car car : this.cars) {
+            if (car.isWinner(max)) {
+                winners.add(car);
+            }
+        }
+        return winners;
     }
 }
